requirements.txt
pytz==2024.2
RPi.bme280==0.2.4
smbus2==0.5.0
